Creating IGraph and abstracting out BaseGraph functions
